# packages/stock-data/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GlobalStockDataService",
    "description": "面向股票分析、量化交易与策略回测的全球股票数据服务 (50GB空间极致优化架构)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/meta/symbols": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "List Symbols",
        "operationId": "list_symbols_api_v1_meta_symbols_get",
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Market"
                },
                {
                  "type": "null"
                }
              ],
              "description": "市场: SH, SZ, US, HK",
              "title": "Market"
            },
            "description": "市场: SH, SZ, US, HK"
          },
          {
            "name": "asset_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssetType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "资产类别: STK, IDX, ETF",
              "title": "Asset Type"
            },
            "description": "资产类别: STK, IDX, ETF"
          },
          {
            "name": "is_benchmark",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否基准资产",
              "title": "Is Benchmark"
            },
            "description": "是否基准资产"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Symbols Api V1 Meta Symbols Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meta/calendar": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "Check Trading Day",
        "operationId": "check_trading_day_api_v1_meta_calendar_get",
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Market",
              "description": "市场"
            },
            "description": "市场"
          },
          {
            "name": "trade_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "日期: YYYY-MM-DD",
              "title": "Trade Date"
            },
            "description": "日期: YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Check Trading Day Api V1 Meta Calendar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/index/constituents": {
      "get": {
        "tags": [
          "Market Intelligence, Screener & Flow"
        ],
        "summary": "Get Index Constituents",
        "description": "获取宽基指数的最新真实成分股列表与股票池（用于策略选股范围限定或指数增强模型）。",
        "operationId": "get_index_constituents_api_v1_index_constituents_get",
        "parameters": [
          {
            "name": "index_symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "指数代码，如 000300 (沪深300), 000905 (中证500), SPX (标普500)",
              "title": "Index Symbol"
            },
            "description": "指数代码，如 000300 (沪深300), 000905 (中证500), SPX (标普500)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstituentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener": {
      "get": {
        "tags": [
          "Market Intelligence, Screener & Flow"
        ],
        "summary": "Run Screener",
        "description": "A 股全市场每日截面选股器 (A-Share Market Screener):\n基于 A 股全市场实时行情截面多因子筛选标的，支持涨跌幅、成交额、价格区间过滤。",
        "operationId": "run_screener_api_v1_screener_get",
        "parameters": [
          {
            "name": "min_pct_change",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "最小涨跌幅(%)，如 3.0 (代表今日涨幅>=3%)",
              "title": "Min Pct Change"
            },
            "description": "最小涨跌幅(%)，如 3.0 (代表今日涨幅>=3%)"
          },
          {
            "name": "max_pct_change",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "最大涨跌幅(%)，如 9.9",
              "title": "Max Pct Change"
            },
            "description": "最大涨跌幅(%)，如 9.9"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "最低价格 (元)",
              "title": "Min Price"
            },
            "description": "最低价格 (元)"
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "最高价格 (元)",
              "title": "Max Price"
            },
            "description": "最高价格 (元)"
          },
          {
            "name": "min_amount",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "最低成交额 (元)，如 100000000 (代表1亿以上)",
              "title": "Min Amount"
            },
            "description": "最低成交额 (元)，如 100000000 (代表1亿以上)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "返回结果上限",
              "default": 50,
              "title": "Limit"
            },
            "description": "返回结果上限"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetType": {
        "type": "string",
        "enum": [
          "STK",
          "IDX",
          "ETF",
          "FX",
          "CRYPTO"
        ],
        "title": "AssetType"
      },
...
      "ConstituentItem": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "description": "规范化标的代码: 如 600519.SH.STK"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "in_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Date",
            "description": "纳入指数日期"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "ticker",
          "name"
        ],
        "title": "ConstituentItem"
      },
      "ConstituentResponse": {
        "properties": {
          "index_symbol": {
            "type": "string",
            "title": "Index Symbol"
          },
          "index_name": {
            "type": "string",
            "title": "Index Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "constituents": {
            "items": {
              "$ref": "#/components/schemas/ConstituentItem"
            },
            "type": "array",
            "title": "Constituents"
          }
        },
        "type": "object",
        "required": [
          "index_symbol",
          "index_name",
          "count",
          "constituents"
        ],
        "title": "ConstituentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Market": {
        "type": "string",
        "enum": [
          "SH",
          "SZ",
          "BJ",
          "US",
          "HK",
          "BINANCE",
          "FX"
        ],
        "title": "Market"
      },
...
      "ScreenerItem": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "latest_price": {
            "type": "number",
            "title": "Latest Price"
          },
          "pct_change": {
            "type": "number",
            "title": "Pct Change"
          },
          "volume": {
            "type": "number",
            "title": "Volume"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "ticker",
          "name",
          "latest_price",
          "pct_change",
          "volume",
          "amount"
        ],
        "title": "ScreenerItem"
      },
      "ScreenerResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ScreenerItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "count",
          "data"
        ],
        "title": "ScreenerResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}